Handle the libssl/libcrypto 0.9.8/1.0.0, rewamp filtering.
We need to handle the deps for libcrypto/libssl. Install script
will resolve the 0.9.8 requires as 1.0.0 iff 1.0.0 is present
and 0.9.8 is *not* present during build. This is too complicated
to fix, so we just add some hard links after 'make install' to
always use 1.0.0 according to mysterious entry in changelog.

Current filtering is broken. It Provides: libcef.so, making it possible
for other apps to link to that library. It also turns off the
dependency generator which makes zypper happily install package
with all sorts of dependency errors unless we state all of them
explicitly (how has this ever worked?).

Removing complete filter solution, using upstream builtin filtering
instead. Requires us to add som manual Requires: for symlinked deps
which not are picked up automatically. No solution is really optimal
here, but this is the best I have found out so far.