install-spotify: download_debs: get all packages, keep previous dl.
In order to generate the builddeps we need a srpm package - there
seem to no equivalent of yum-builddep which can get the buildeps
for a spec file in the zypper universe.

Building a srpm means that both packages unfortunately are needed,if
we want to stick to rpm semantics.

Keeping the old packages in the dl area means that they will be reused
onlater builds.